golinks internal redirect tools
This commit introduces the service that runs at go.corp.mixpanel.com.

In essence, go/ allows you to define small shortcuts that redirects
users to expanded URLs. They can be created by any one in the company
and is available for use by everyone in the company. If you a notice a
broken go link, you can fix it even if you didn't originally create it!

Change-Id: Ibc670cc340743d9f9215f32a5290b4bb67dd7b05