Change go/src/logger to log all Fields locally by default
Instead of relying on {{template_var}} style substitution, make the
logger log all fields in a standard way. This has couple of benefits:

(1) Writing log lines becomes a bit easier as the redundancy is reduced
(2) We can easily add duplicate line detection pretty easily as the main
message string won't change. NOTE: this is not implemented in this
review.

Change-Id: I6823352169eca0e8f6b32a81981781f28b8d9d3d