Fetch and print hostname for a given pod name
- Add default tags to wavefront sink
- Fetch and print hostname for a given pod's nodename

The kubernetes folks are already working on a PR (https://github.com/kubernetes/kubernetes/pull/26160)
to propagate the pod spec using an environment variable. Till that is available, we will just use
this PR to fetch the hostname for a given pod and propagate that to statsd.

Tested on golinks.

Change-Id: I0da01fb9811dfb88d13224e1b5687aafe354df84