arb3: make tracing to StackDriver Trace work.
This add some utilities to httputil and obs to help with instrumenting
gRPC and HTTP client/servers. It also switches the LQS to use FlightRecorder.

The contexts aren't properly threaded through for the LQS because we
have some long-lived requests but I have an idea on how to make it
work so I'll follow up with another branch that improves the LQS.

Change-Id: Ib8eb3c580ec10b2248290656e4526d9169db0606